+++ cli/replicaset/lua/cconfig/get_sharding_roles_body.lua
@@ -1,0 +1,5 @@
+local sharding = require('config'):get().sharding
+if sharding == nil or sharding.roles == nil then
+    error("sharding roles are not configured, please make sure managed cluster is sharded")
+end
+return sharding.roles